Wired through blog page connection to prod
